# Support udecimal library in code generation