--- a/cmd/generate-fix/internal/template_helpers.go
+++ b/cmd/generate-fix/internal/template_helpers.go
@@ -6,6 +6,15 @@ import (
 	"github.com/quickfixgo/quickfix/datadictionary"
 )
 
+func isDecimalType(quickfixType string) bool {
+	switch quickfixType {
+	case "FIXDecimal", "FIXUDecimal":
+		return true
+	default:
+		return false
+	}
+}
+
 func checkIfDecimalImportRequiredForFields(fTypes []*datadictionary.FieldType) (ok bool, err error) {
 	var t string
 	for _, fType := range fTypes {
@@ -14,7 +23,7 @@ func checkIfDecimalImportRequiredForFields(fTypes []*datadictionary.FieldType) (
 			return
 		}
 
-		if t == "FIXDecimal" {
+		if isDecimalType(t) {
 			return true, nil
 		}
 	}
@@ -54,7 +63,7 @@ func checkFieldDecimalRequired(f *datadictionary.FieldDef) (required bool, err e
 		return
 	}
 
-	if t == "FIXDecimal" {
+	if isDecimalType(t) {
 		required = true
 		return
 	}
@@ -121,6 +130,10 @@ func collectStandardImports(m *datadictionary.MessageDef) (imports []string, err
 
 func collectExtraImports(m *datadictionary.MessageDef) (imports []string, err error) {
 	var decimalRequired bool
+	importPath := "github.com/shopspring/decimal"
+	if *useUDecimal {
+		importPath = "github.com/quagmt/udecimal"
+	}
 	for _, f := range m.Fields {
 		if !decimalRequired {
 			if decimalRequired, err = checkFieldDecimalRequired(f); err != nil {
@@ -134,7 +147,7 @@ func collectExtraImports(m *datadictionary.MessageDef) (imports []string, err er
 	}
 
 	if decimalRequired {
-		imports = append(imports, "github.com/shopspring/decimal")
+		imports = append(imports, importPath)
 	}
 
 	return
@@ -191,6 +204,8 @@ func quickfixValueType(quickfixType string) (goType string, err error) {
 		goType = "float64"
 	case "FIXDecimal":
 		goType = "decimal.Decimal"
+	case "FIXUDecimal":
+		goType = "udecimal.Decimal"
 	default:
 		err = fmt.Errorf("Unknown QuickFIX Type: %v", quickfixType)
 	}
@@ -275,6 +290,8 @@ func quickfixType(field *datadictionary.FieldType) (quickfixType string, err err
 	case "FLOAT":
 		if *useFloat {
 			quickfixType = "FIXFloat"
+		} else if *useUDecimal {
+			quickfixType = "FIXUDecimal"
 		} else {
 			quickfixType = "FIXDecimal"
 		}